updated with latest code
